refactor: factorize result component
